refactor: migrate remaining imports to shared.* and fix test mock paths (#2413)
Phase 2b completion: Update all remaining deprecated imports and test mocks.

Changes:
- Batch migrate source imports from path_helpers to shared.paths
- Fix cross-module import in run_helpers to use public API (lazy import)
- Update test mock patches to use public API paths
- Update test imports for shared.paths module
- Fix comment references to new module paths

Architecture compliance:
- Cross-module imports use public API (vibe3.services.*)
- Tests patch public API symbols for mock injection
- Module-specific imports use direct paths (shared.paths)

Tests: All 3020 tests pass (4 xfailed, 1 xpassed)
Types: mypy strict passes (413 source files)

Follows: #2411 (Phase 1 architecture fix)
Related: #2392 (original migration PR)
