+++ src/vibe3/services/handoff_service.py
@@ -14,10 +14,10 @@ from vibe3.services.actor_support import (
 )
 from vibe3.services.artifact_parser import ArtifactParser
 from vibe3.services.external_events import ExternalEventRecorder
-from vibe3.services.git_path_client import GitPathProtocol
+from vibe3.services.handoff_resolution import _SHARED_HANDOFF_PREFIX
 from vibe3.services.handoff_storage import HandoffStorage
 from vibe3.services.handoff_validation import validate_authoritative_ref
-from vibe3.services.path_helpers import _SHARED_HANDOFF_PREFIX
+from vibe3.services.shared.paths import GitPathProtocol
 from vibe3.services.signature_service import SignatureService
 
 if TYPE_CHECKING:
