+++ src/vibe3/services/task/classifier.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from enum import Enum
 
 from vibe3.models import IssueState
-from vibe3.services.label_utils import has_manager_assignee
+from vibe3.services.shared.labels import has_manager_assignee
 
 
 class TaskStatusBucket(str, Enum):
